refactor: update to work with latest pyo3
